feat: removed clipboard api from @effect/platform
